feat: intersection observer visualizer와 서비스를 연결할 수 있도록 추가
