+++ src/manager/left-navigation-bar.tsx
@@ -11,6 +11,7 @@ import { TabWidget } from './tabs/widget';
 import { Formik, FormikHelpers, FormikProvider, useFormik } from 'formik';
 import { SELECT_UNIT_LIST } from './observer-option-form/constants';
 import { emit } from '../util/custom-event';
+import { useEffect } from 'react';
 
 const Container = styled.div<{ isOpen: boolean }>`
   width: 250px;
@@ -71,6 +72,17 @@ const initialValues = {
   },
 };
 
+interface CreateRectArgs {
+  x: number;
+  y: number;
+  bottom: number;
+  color: string;
+  height: number;
+  left: number;
+  right: number;
+  top: number;
+  width: number;
+}
 export type IntersectionObserverOptionFormType = typeof initialValues;
 
 export function LeftNavigationBar() {
@@ -84,6 +96,107 @@ export function LeftNavigationBar() {
       emit(window, '@submit', values);
     },
   });
+
+  useEffect(() => {
+    const createActiveRectangle = ({
+      x,
+      y,
+      bottom,
+      color,
+      height,
+      left,
+      right,
+      top,
+      width,
+    }: CreateRectArgs) => {
+      const div = document.createElement('div');
+
+      div.style.position = 'absolute';
+      div.style.left = `${x - 1}px`;
+      div.style.top = `${y - 1}px`;
+      div.style.width = `${width + 1}px`;
+      div.style.height = `${height + 1}px`;
+      div.style.border = `2px solid ${color}`;
+      div.style.zIndex = '1';
+      div.style.marginLeft = `${left}px`;
+      div.style.marginTop = `${top}px`;
+      div.style.marginRight = `${right - width}px`;
+      div.style.marginBottom = `${bottom - height}px`;
+      div.style.pointerEvents = 'none'; // 이렇게 할 시 네모에 상호작용 안됨.
+      // div.style.transform = 'scale(0.75)';
+      div.style.transition = 'opacity 100s ease-in';
+
+      console.log({ left });
+      console.log({ top });
+      console.log({ right: right - width });
+      console.log({ bottom: bottom - height });
+
+      requestAnimationFrame(() =>
+        requestAnimationFrame(() => {
+          div.style.opacity = '0';
+        }),
+      );
+
+      div.addEventListener('transitionend', () => {
+        document.body.removeChild(div);
+      });
+
+      document.body.append(div);
+    };
+
+    const drawRectByMessage = () => {
+      window.addEventListener('message', (e) => {
+        const entry = typeof e.data === 'string' ? JSON.parse(e.data) : e.data;
+
+        if (!entry || !Object.keys(entry).some((key) => key === 'rootBounds')) {
+          return;
+        }
+
+        const rootBounds = entry.rootBounds;
+        const boundingClientRect = entry.boundingClientRect;
+        const $iframe = document.querySelector('.io-iframe');
+        const rect = $iframe?.getBoundingClientRect() as DOMRect;
+
+        const x = rect.left + window.scrollX - 1;
+        const y = rect.top + window.scrollY - 1;
+
+        console.log({ rect });
+        console.log({ boundingClientRect });
+        console.log({ rootBounds });
+
+        if (!rootBounds) {
+          console.log('rootBounds가 없습니다.');
+          return;
+        }
+
+        createActiveRectangle({
+          x,
+          y,
+          bottom: 0.75 * rootBounds.bottom,
+          color: 'red',
+          height: 0.75 * rootBounds.height,
+          left: 0.75 * rootBounds.left,
+          right: 0.75 * rootBounds.right,
+          top: 0.75 * rootBounds.top,
+          width: 0.75 * rootBounds.width,
+        });
+
+        createActiveRectangle({
+          x,
+          y,
+          bottom: 0.75 * boundingClientRect.bottom,
+          color: 'blue',
+          height: 0.75 * boundingClientRect.height,
+          left: 0.75 * boundingClientRect.left,
+          right: 0.75 * boundingClientRect.right,
+          top: 0.75 * boundingClientRect.top,
+          width: 0.75 * boundingClientRect.width,
+        });
+      });
+    };
+
+    drawRectByMessage();
+  }, []);
 
   return (
     <Container isOpen={isOpen}>
